pyproject.toml
[tool.poetry]
name = "smart-code-planner"
version = "0.1.0"
description = "A sophisticated code organization agent built with LangGraph"
authors = ["Federico Antosiano <federico.antosiano@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.0"
langgraph = "^0.2.0"
langchain-openai = "^0.2.0"
langchain-google-genai = "^2.1.0"
streamlit = "^1.32.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
python-dotenv = "^1.0.0"
typing-extensions = "^4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.0.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
